Fix MLX compilation int64/int32 data type issues
- Fix tokenizer cache to use int32 instead of int64 for MLX arrays
- Update both tokenizer_cache.py and tokenizer_cache_mlx.py
- Clear existing cache with incompatible int64 data types
- Enable successful MLX compilation with proper data types
- Fix KaggleTrainer tokenizer initialization
- Fix dataset label extraction for cross-validation
- Complete data pipeline fixes for Titanic competition training

+++ data/preprocessing/tokenizer_cache.py
@@ -109,15 +109,15 @@ class TokenizerCache:
         )
 
         # Convert to MLX arrays (unified memory)
-        # Note: Tokenizer returns numpy arrays, convert to MLX
+        # Note: Tokenizer returns numpy arrays, convert to MLX with int32 dtype for compilation compatibility
         tokenized_data = {
-            "input_ids": mx.array(encoded["input_ids"]),
-            "attention_mask": mx.array(encoded["attention_mask"]),
+            "input_ids": mx.array(encoded["input_ids"], dtype=mx.int32),
+            "attention_mask": mx.array(encoded["attention_mask"], dtype=mx.int32),
         }
 
         # Add token_type_ids if available
         if "token_type_ids" in encoded:
-            tokenized_data["token_type_ids"] = mx.array(encoded["token_type_ids"])
+            tokenized_data["token_type_ids"] = mx.array(encoded["token_type_ids"], dtype=mx.int32)
 
         # Add labels if provided
         if labels is not None:
